added prettier. refactored dataverseservice
