+++ PolyLookupComponent/.eslintrc.json
@@ -1,30 +1,39 @@
 {
-    "env": {
-      "browser": true,
-      "es2021": true
-    },
-    "extends": [
-      "eslint:recommended",
-      "plugin:react/recommended"
-    ],
-    "globals": {
-      "ComponentFramework": true
-    },
-    "parser": "@typescript-eslint/parser",
-    "parserOptions": {
-      "ecmaVersion": 12,
-      "sourceType": "module"
-    },
-    "plugins": [
-      "@microsoft/power-apps",
-      "@typescript-eslint"
-    ],
-    "rules": {
-      "no-unused-vars": "off"
-    },
-    "settings": {
-      "react": {
-        "version": "detect"
+  "env": {
+    "browser": true,
+    "es2021": true
+  },
+  "extends": [
+    "plugin:@typescript-eslint/recommended",
+    "plugin:prettier/recommended",
+    "plugin:react/recommended",
+    "prettier"
+  ],
+  "globals": {
+    "ComponentFramework": true
+  },
+  "parser": "@typescript-eslint/parser",
+  "parserOptions": {
+    "ecmaVersion": 12,
+    "sourceType": "module"
+  },
+  "plugins": ["@microsoft/power-apps", "@typescript-eslint", "prettier"],
+  "rules": {
+    "no-unused-vars": "off",
+    "no-explicit-any": "off",
+    "prettier/prettier": "warn"
+  },
+  "settings": {
+    "react": {
+      "version": "detect"
+    }
+  },
+  "overrides": [
+    {
+      "files": ["*.ts"],
+      "rules": {
+        "camelcase": [2, { "properties": "never" }]
       }
     }
+  ]
 }
